outbound-governor: LedgerPeers module
It is confusing to have non DNS functionality under
`RootPeersDNS.LedgerPeers` and hence most of the module is put back in
`PeerSelection.LedgerPeers`, only the `resolveLedgerPeers` function is
left in the `RootPeersDNS.LedgerPeers` module.  To make that happen we
also need `PeerSelection.LedgerPeers.Common`.
