+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/PeerSelection/Instances.hs
@@ -20,13 +20,12 @@ import           Data.Word (Word32)
 import           Ouroboros.Network.PeerSelection.Governor
 
 import qualified Data.IP as IP
+import           Ouroboros.Network.PeerSelection.LedgerPeers (IsLedgerPeer (..))
 import           Ouroboros.Network.PeerSelection.PeerAdvertise
                      (PeerAdvertise (..))
 import           Ouroboros.Network.PeerSelection.PeerSharing (PeerSharing (..))
 import           Ouroboros.Network.PeerSelection.RelayAccessPoint
                      (DomainAccessPoint (..), RelayAccessPoint (..))
-import           Ouroboros.Network.PeerSelection.RootPeersDNS.LedgerPeers
-                     (IsLedgerPeer (..))
 import           Ouroboros.Network.Testing.Utils (ShrinkCarefully,
                      prop_shrink_nonequal, prop_shrink_valid)
 import           Test.QuickCheck
