+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/PeerSelection/KnownPeers.hs
@@ -1,10 +1,10 @@
 
 module Test.Ouroboros.Network.PeerSelection.KnownPeers (tests) where
 import           Data.Map (Map)
+import           Ouroboros.Network.PeerSelection.LedgerPeers
 import           Ouroboros.Network.PeerSelection.PeerAdvertise (PeerAdvertise)
 import           Ouroboros.Network.PeerSelection.PeerSharing (PeerSharing)
 import           Ouroboros.Network.PeerSelection.RelayAccessPoint
-import           Ouroboros.Network.PeerSelection.RootPeersDNS.LedgerPeers
 import qualified Ouroboros.Network.PeerSelection.State.KnownPeers as KnownPeers
 import           Test.Ouroboros.Network.PeerSelection.Instances ()
 import           Test.QuickCheck (Property, counterexample)
