+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/LedgerPeers/Common.hs
@@ -1,0 +1,104 @@
+-- | Types used by `Ouroboros.Network.PeerSelection.LedgerPeers` and
+-- `Ouroboros.Network.PeerSelection.RootPeersDNS.LedgerPeers`
+--
+module Ouroboros.Network.PeerSelection.LedgerPeers.Common where
+
+import           Control.Monad.Class.MonadTime.SI
+import           Data.Word (Word16)
+import           Text.Printf
+
+import qualified Network.DNS as DNS
+
+import           Cardano.Slotting.Slot (SlotNo)
+
+import           Ouroboros.Network.PeerSelection.LedgerPeers.Type
+import           Ouroboros.Network.PeerSelection.RelayAccessPoint
+
+-- | Only use the ledger after the given slot number.
+data UseLedgerAfter = DontUseLedger | UseLedgerAfter SlotNo deriving (Eq, Show)
+
+isLedgerPeersEnabled :: UseLedgerAfter -> Bool
+isLedgerPeersEnabled DontUseLedger = False
+isLedgerPeersEnabled _             = True
+
+newtype NumberOfPeers = NumberOfPeers { getNumberOfPeers :: Word16 }
+  deriving Show
+
+-- | Identifies a peer as coming from ledger or not
+data IsLedgerPeer = IsLedgerPeer
+                  -- ^ a ledger peer.
+                  | IsNotLedgerPeer
+  deriving (Eq, Show)
+
+-- | Trace LedgerPeers events.
+data TraceLedgerPeers =
+      PickedBigLedgerPeer RelayAccessPoint AccPoolStake PoolStake
+      -- ^ Trace for a significant ledger peer picked with accumulated and relative stake of its pool.
+    | PickedLedgerPeer RelayAccessPoint AccPoolStake PoolStake
+      -- ^ Trace for a ledger peer picked with accumulated and relative stake of its pool.
+    | PickedBigLedgerPeers NumberOfPeers [RelayAccessPoint]
+    | PickedLedgerPeers    NumberOfPeers [RelayAccessPoint]
+      -- ^ Trace for the number of peers and we wanted to pick and the list of peers picked.
+    | FetchingNewLedgerState Int Int
+      -- ^ Trace for fetching a new list of peers from the ledger. The first Int
+      -- is the number of ledger peers returned the latter is the number of big
+      -- ledger peers.
+    | TraceLedgerPeersDomains [DomainAccessPoint]
+    | TraceLedgerPeersResult  DNS.Domain [(IP, DNS.TTL)]
+    | TraceLedgerPeersFailure DNS.Domain DNS.DNSError
+    | DisabledLedgerPeers
+      -- ^ Trace for when getting peers from the ledger is disabled, that is DontUseLedger.
+    | TraceUseLedgerAfter UseLedgerAfter
+      -- ^ Trace UseLedgerAfter value
+    | WaitingOnRequest
+    | RequestForPeers NumberOfPeers
+    | ReusingLedgerState Int DiffTime
+    | FallingBackToPublicRootPeers
+    | NotEnoughBigLedgerPeers NumberOfPeers Int
+    | NotEnoughLedgerPeers NumberOfPeers Int
+
+
+instance Show TraceLedgerPeers where
+    show (PickedBigLedgerPeer addr ackStake stake) =
+        printf "PickedBigLedgerPeer %s ack stake %s ( %.04f) relative stake %s ( %.04f )"
+            (show addr)
+            (show $ unAccPoolStake ackStake)
+            (fromRational (unAccPoolStake ackStake) :: Double)
+            (show $ unPoolStake stake)
+            (fromRational (unPoolStake stake) :: Double)
+    show (PickedLedgerPeer addr ackStake stake) =
+        printf "PickedLedgerPeer %s ack stake %s ( %.04f) relative stake %s ( %.04f )"
+            (show addr)
+            (show $ unAccPoolStake ackStake)
+            (fromRational (unAccPoolStake ackStake) :: Double)
+            (show $ unPoolStake stake)
+            (fromRational (unPoolStake stake) :: Double)
+    show (PickedBigLedgerPeers (NumberOfPeers n) peers) =
+        printf "PickedBigLedgerPeers %d %s" n (show peers)
+    show (PickedLedgerPeers (NumberOfPeers n) peers) =
+        printf "PickedLedgerPeers %d %s" n (show peers)
+    show (FetchingNewLedgerState cnt bigCnt) =
+        printf "Fetching new ledgerstate, %d registered pools, %d registered big ledger pools"
+            cnt bigCnt
+    show (TraceUseLedgerAfter ula) =
+        printf "UseLedgerAfter state %s"
+            (show ula)
+    show WaitingOnRequest = "WaitingOnRequest"
+    show (RequestForPeers (NumberOfPeers cnt)) = printf "RequestForPeers %d" cnt
+    show (ReusingLedgerState cnt age) =
+        printf "ReusingLedgerState %d peers age %s"
+          cnt
+          (show age)
+    show FallingBackToPublicRootPeers = "Falling back to public root peers"
+    show DisabledLedgerPeers = "LedgerPeers is disabled"
+    show (NotEnoughBigLedgerPeers (NumberOfPeers n) numOfBigLedgerPeers) =
+      printf "Not enough big ledger peers to pick %d out of %d" n numOfBigLedgerPeers
+    show (NotEnoughLedgerPeers (NumberOfPeers n) numOfLedgerPeers) =
+      printf "Not enough ledger peers to pick %d out of %d" n numOfLedgerPeers
+
+    show (TraceLedgerPeersDomains domains) = "Resolving " ++ show domains
+    show (TraceLedgerPeersResult domain l) =
+      "Resolution success " ++ show domain ++ " " ++ show l
+    show (TraceLedgerPeersFailure domain err) =
+      "Resolution failed " ++ show domain ++ " " ++ show err
+
